requirements.txt
python-dotenv
streamlit
unstructured
uuid
pathlib
chromadb
langchain_groq
langchain_community
langchain